# replace ToUtf8 with Encoding.UTF8.GetBytes