--- a/src/libraries/System.Text.Json/src/System/Text/Json/Reader/Utf8JsonReader.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Reader/Utf8JsonReader.cs
@@ -531,21 +531,8 @@ namespace System.Text.Json
                 otherUtf8Text = stackalloc byte[JsonConstants.StackallocByteThreshold];
             }
 
-            ReadOnlySpan<byte> utf16Text = MemoryMarshal.AsBytes(text);
-            OperationStatus status = JsonWriterHelper.ToUtf8(utf16Text, otherUtf8Text, out int consumed, out int written);
-            Debug.Assert(status != OperationStatus.DestinationTooSmall);
-            bool result;
-            if (status > OperationStatus.DestinationTooSmall)   // Equivalent to: (status == NeedMoreData || status == InvalidData)
-            {
-                result = false;
-            }
-            else
-            {
-                Debug.Assert(status == OperationStatus.Done);
-                Debug.Assert(consumed == utf16Text.Length);
-
-                result = TextEqualsHelper(otherUtf8Text.Slice(0, written));
-            }
+            int written = Encoding.UTF8.GetBytes(text, otherUtf8Text);
+            bool result = TextEqualsHelper(otherUtf8Text.Slice(0, written));
 
             if (otherUtf8TextArray != null)
             {