--- a/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Bytes.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Bytes.cs
@@ -231,7 +231,8 @@ namespace System.Text.Json
             }
             output[BytesPending++] = JsonConstants.Quote;
 
-            BytesPending += Encoding.UTF8.GetBytes(escapedPropertyName, output.Slice(BytesPending));
+            JsonWriterHelper.TryGetUtf8FromText(escapedPropertyName, output.Slice(BytesPending), out int written);
+            BytesPending += written;
 
             output[BytesPending++] = JsonConstants.Quote;
             output[BytesPending++] = JsonConstants.KeyValueSeparator;
@@ -316,7 +317,8 @@ namespace System.Text.Json
 
             output[BytesPending++] = JsonConstants.Quote;
 
-            BytesPending += Encoding.UTF8.GetBytes(escapedPropertyName, output.Slice(BytesPending));
+            JsonWriterHelper.TryGetUtf8FromText(escapedPropertyName, output.Slice(BytesPending), out int written);
+            BytesPending += written;
 
             output[BytesPending++] = JsonConstants.Quote;
             output[BytesPending++] = JsonConstants.KeyValueSeparator;