--- a/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Double.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Double.cs
@@ -237,7 +237,7 @@ namespace System.Text.Json
             }
             output[BytesPending++] = JsonConstants.Quote;
 
-            TranscodeAndWrite(escapedPropertyName, output);
+            BytesPending += Encoding.UTF8.GetBytes(escapedPropertyName, output.Slice(BytesPending));
 
             output[BytesPending++] = JsonConstants.Quote;
             output[BytesPending++] = JsonConstants.KeyValueSeparator;
@@ -313,7 +313,7 @@ namespace System.Text.Json
 
             output[BytesPending++] = JsonConstants.Quote;
 
-            TranscodeAndWrite(escapedPropertyName, output);
+            BytesPending += Encoding.UTF8.GetBytes(escapedPropertyName, output.Slice(BytesPending));
 
             output[BytesPending++] = JsonConstants.Quote;
             output[BytesPending++] = JsonConstants.KeyValueSeparator;