@eiriktsarpalis 2022-09-19
> Can we validate performance is as we expect it to be (as good or better than it was before)?

@kasperk81 If you're interested, here are [a few instructions](https://github.com/dotnet/performance/blob/main/src/benchmarks/micro/README.md#private-runtime-builds) on how to benchmark private build performance. The STJ benchmarks can be run using the `System.Text.Json*` filter in the BenchmarkDotNet. I can also do it for you if you prefer.