@kasperk81 2022-09-19
> Can we validate performance is as we expect it to be (as good or better than it was before)?

from one set of performance tests, i'm getting mixed results.

summary:
better: 5, geomean: 1.087
worse: 4, geomean: 1.115
total diff: 9

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.16 |           754.87 |           876.72 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.15 |           863.18 |           992.44 | bimodal |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.12 |       1993853.75 |       222316.66 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.04 |       1911530.87 |       1982514.57 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.15 |           832.79 |           721.12 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.08 |          1063.42 |           980.44 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.08 |          1019.21 |           940.08 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: True, SkipValidation |      1.07 |          1003.61 |           941.22 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.05 |           710.38 |           676.76 |         |