@EgorBo 2022-09-19
@kasperk81 I'd recommend to pass `--envVars DOTNET_ReadyToRun:0` to BDN to disable static PGO as your changes changed layout and made the built-in profile stale.
I assume we need to reflect it in the docs (https://github.com/dotnet/performance/blob/main/src/benchmarks/micro/README.md#private-runtime-builds)