@kasperk81 2022-09-19
> I'd recommend to pass `--envVars DOTNET_ReadyToRun:0`

with envVars, it's clear that diff is slower
```
summary:
worse: 10, geomean: 1.132
total diff: 10

| Slower                                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.29 |       2362864.09 |       3058230.01 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: True, SkipValidation |      1.22 |       2414225.03 |       2944031.19 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.21 |           689.49 |           834.19 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.14 |           748.64 |           856.18 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.12 |           963.79 |          1079.45 | bimodal |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: True, SkipValidatio |      1.10 |           908.93 |           995.58 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf16(Formatted: False, SkipValidati |      1.07 |       1838042.24 |       1969624.88 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.07 |           634.40 |           677.76 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: True, SkipValidation |      1.06 |           893.44 |           947.61 |         |
| System.Text.Json.Tests.Perf_Basic.WriteBasicUtf8(Formatted: False, SkipValidatio |      1.06 |           707.75 |           750.28 |         |

No Faster results for the provided threshold = 2% and noise filter = 0.3 ns.
```