@eiriktsarpalis 2022-10-24
I've analyzed the test failures and made the following discoveries. Unlike other strings passed to `Utf8JsonWriter`, comment strings are not escaped, by design:

https://github.com/dotnet/runtime/blob/58a1180ad6421c6f69cf307b57b2496c888585f2/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteValues.Comment.cs#L45

As a result, the comment writer supports inputs containing unpaired surrogates:

https://github.com/dotnet/runtime/blob/58a1180ad6421c6f69cf307b57b2496c888585f2/src/libraries/System.Text.Json/tests/System.Text.Json.Tests/Utf8JsonWriterTests.cs#L4165

However, in such cases the current implementation will simply write the comment string up to the impacted character and discard everything else

https://github.com/dotnet/runtime/blob/58a1180ad6421c6f69cf307b57b2496c888585f2/src/libraries/System.Text.Json/tests/System.Text.Json.Tests/Utf8JsonWriterTests.cs#L4200

Even though we're testing for this, I don't think anybody could reasonably depend on the current behavior. I would recommend taking a breaking change and have the method throw ArgumentException instead.